+++ src/rtcmix/PField.h
@@ -23,7 +23,7 @@ public:
 #endif
 	virtual double 	doubleValue(int indx = 0) const = 0;
 	virtual double 	doubleValue(double dindex) const = 0;
-	int 			intValue(double dindex) const { return (int) doubleValue(dindex); }
+	size_t 			intValue(double dindex) const { return (size_t) doubleValue(dindex); }
 	const char * 	stringValue(double dindex) const;
 	virtual int		print(FILE *) const;
 	virtual operator double *() const { /* default is to */ return 0; }
